+++ packages/interactor/src/index.ts
@@ -4,6 +4,7 @@ export { createInteractor } from './create-interactor';
 export { createInspector } from './inspector'
 export { Page } from './page';
 export { App } from './app';
+export { read } from './read';
 export { perform } from './perform';
 export { fillIn } from './fill-in';
 export { focused, focus, blur } from './focused';