+++ packages/interactor/src/specification.ts
@@ -178,6 +178,10 @@ export type ActionMethods<E extends Element, A extends Actions<E>> = {
     : never;
 }
 
+export type FilterReturn<F> = {
+  [P in keyof F]?: F[P] extends MaybeMatcher<infer T> ? T : never;
+}
+
 export type FilterParams<E extends Element, F extends Filters<E>> = keyof F extends never ? never : {
   [P in keyof F]?:
     F[P] extends FilterFn<infer TArg, E> ?